replace JavaPrinter and TreePrinter
