--- rewrite-core/src/main/java/org/openrewrite/TreePrinter2.java
@@ -1,33 +1,0 @@
-package org.openrewrite;
-
-import org.openrewrite.internal.lang.Nullable;
-
-public interface TreePrinter2<P> {
-
-    static <P> TreePrinter2<P> identity() {
-        return new TreePrinter2<P>() {
-        };
-    }
-
-    /**
-     * Called before tree is printed, allows printing additional output before tree.
-     *
-     * @param tree AST element that is about to be printed
-     * @param printerAcc Printer accumulator, can be appended to, contains everything written for the overall visit
-     *                   operation so far
-     * @param p visit context
-     */
-    default void doBefore(@Nullable Tree tree, StringBuilder printerAcc, P p) {}
-
-    /**
-     * Called after tree has been printed, allows printing additional output after tree.
-     *
-     * @param tree AST element that has just been printed
-     * @param printerAcc Printer accumulator, can be appended to, contains everything written for the overall visit
-     *      *                   operation so far
-     * @param p visit context
-     */
-    default void doAfter(@Nullable Tree tree, StringBuilder printerAcc, P p) {}
-
-
-}
