+++ rewrite-java/src/main/java/org/openrewrite/java/format/BlankLinesProcessor.java
@@ -16,7 +16,6 @@
 package org.openrewrite.java.format;
 
 import org.openrewrite.Cursor;
-import org.openrewrite.Tree;
 import org.openrewrite.internal.ListUtils;
 import org.openrewrite.java.JavaIsoProcessor;
 import org.openrewrite.java.style.BlankLinesStyle;
