+++ rewrite-test/src/main/kotlin/org/openrewrite/java/AddImportTest.kt
@@ -18,7 +18,6 @@ package org.openrewrite.java
 import org.junit.jupiter.api.Test
 import org.openrewrite.*
 import org.openrewrite.java.tree.J
-import java.util.function.Supplier
 
 interface AddImportTest : RecipeTest {
     fun addImports(vararg adds: AddImport<ExecutionContext>): Recipe = adds
