+++ rewrite-test/src/main/kotlin/org/openrewrite/java/JavaProcessorCompatibilityKit.kt
