+++ rewrite-test/src/main/kotlin/org/openrewrite/java/UnwrapParenthesesTest.kt
@@ -21,7 +21,6 @@ import org.openrewrite.Recipe
 import org.openrewrite.RecipeTest
 import org.openrewrite.TreeProcessor
 import org.openrewrite.java.tree.J
-import java.util.function.Supplier
 
 interface UnwrapParenthesesTest : RecipeTest {
     override val recipe: Recipe?
